backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
pdfplumber
python-docx
werkzeug
